add uncrustify code formatter support
This adds configuration file and script for running the uncrustify
source code beautifier to get a clean and consistent source code
formatting.

The configuration file is selected to generate the minimal diff for now.

As the project can be attributed to the RAUC project, the uncrustify
config also documents the current difference in coding style from RAUC.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>