src/config-file: fix error handling in get_key_bool()
If the value for a boolean key could not be interpredet get_key_bool()
returned FALSE with GError unset.
This violates GError handling rules but will break callers expecting to
be able printing the error message.

To fix this set an error value similar to the one that glib would emit
in g_key_file_get_boolean() when running in the same case.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>