src/config-file: fix error paths in get_key_int()
The handing in get_key_int() had 2 major issues:

* A value of '0' in the config file always resulted in the default value.
  There might be cases where this is a desired behavior, but this should
  not be handled in the generic configuration file parsing as this hides
  configuration errors.

* A fully invalid configuration file integer value like 'zero' did not
  cause any error but only let to loading the default value.
  This may hide severe configuration issues.

Fixed this behavior by a) explicitly checking for a non-set configuration
value and setting the default value then and b) checking for value '0'
with error set to differentiate between generic errors and the parsed
value '0'.

It is required to use an internal GError to allow freeing it in case it
is not propagated.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>