.travis.yml: prepare hawkbit docker to be running for tests
This pulls and runs the hawkbit-update-server docker container to have a
test instance of hawkbit we can use to run unit tests again.

As we need to be sure hawkbit is fully up and running before we can
continue testing, a little script that waits for hawkbit to accept
requests was added, too.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>