src: let hawkbit_start_service_sync() return an exit code
The call of rauc-hawkbit-updater always results in a zero exit code if
the initial command line parameter checks are passed, even if there were
severe failures inside the actual hawkbit-client code.

This happens, because the hawkbit-client code is not able to propagate
error codes to the upper layers.

We start fixing this by adding a return code to the prior void function
hawkbit_start_service_sync() and directly returning the result of
the hawkbit_start_service_sync() call as the program exit code.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>