test: add simple test that identifies with hawkbit and checks for update
To satisfy the authentication model of hawkBit, we must manually add the
target to register with using the hawkbit MGMT API.

Only if both the pre-registered target and the registering client report
the same security token, we are allowed to push device information and
poll for updates.

This adds two tests: One that identifies with a valid security token and
thus must be able to poll for updates, and another one with an invalid
security token that hawkbit must reject.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>