src/config-file: fix error paths in get_groups()
First of all we need to handle the g_key_file_get_keys() return value
that may indicate the group was not found. We can safely abort here and return
'FALSE' with GError set.

Checking if the group does not contain any keys should be done before
iterating over it, thus this check was moved before the iteration.
Additionally, we must set the GError here when returning 'FALSE'.

Last but not least me should also care for potential errors of
g_key_file_get_value() and check and return 'FALSE' here too if
required.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>