src/hawkbit-client: set error in case of fopen or curl init failure
When fopen() fails only a debug message is generated and no GError is
set. This replaces the g_debug() output by setting of GError to allow
propagating the error cause to the caller and to consistently set GError
on failure.

Also if curl_easy_init() fails, we must should set an appropriate error if we
return with an error condition.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>