src/hawkbit-client: split return and http code in get_binary()
Prior to this, error and return code handling in get_binary was ..
The function returned:
* negative numbers if fopen or curl initialization fails
* http error codes
* http success codes (2xx)

This makes it very hard to clearly determine when the function did
actually return successfully and when it failed.

Additionally, the GError was set if curl_easy_perform() did not report
'CURLE_OK', which potentially represents failures independent from the
actual http error code that is then used as return value.

To fix and simplify this handling, we split up the return code of
get_binary to only indicate actual errors or not and return the result
of the HTTP call, i.e. the HTTP status code as an output function
parameter.

This finally allows download_thread() to safely access the GError
information returned by get_binary() and report it to the user.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>