.travis.yml: update dist to ubuntu focal (20.04)
This is useful anyway as it provides much more recent packages.
But the primary reason for this is that gdbus-codegen contained in
ubuntu xenial (16.04) generates C code that triggers build warnings
(that now became errors when building with -Werror).

This was fixed in glib commit f952fdf3 that appeared in 2.53.1 first
while xenial contains 2.48.2

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>