hawkbit-client/rauc-hawkbit-updater: simplify sleep times
The sleep mechanism used is overly complex. Various variables are used:
  - DEFAULT_SLEEP_TIME_SEC used if HawkBit does not provide a sleep
    time (which is not specified in the DDI API docs)
  - hawkbit_config->retry_wait used on HTTP errors
  - hawkbit_interval_check_sec used if HawkBit provided a sleep time
  - sleep_time set to one of the above
  - last_run_sec at last is incremented and checked against sleep_time

Simplify this by:
- getting rid of DEFAULT_SLEEP_TIME_SEC and use
  hawkbit_config->retry_wait instead
- getting rid of sleep_time, use hawkbit_interval_check_sec directly
  instead
- moving away from global variables, instead moving them into
  the ClientData struct

Additionally instead of waiting for the sleep time to pass on the first
run, set the elapsed time (last_run_sec) to the sleep time, so the very
first run already triggers a check. This helps debugging and makes the
force_check_run special treatment obsolete.

Signed-off-by: Bastian Krause <bst@pengutronix.de>