hawkbit-client: remove unneeded get_binary_checksum struct
The struct consists of the actual checksum result and the checksum type.
As the checksum type is hardcoded to sha1, there is no benefit from
passing this along in a struct.

For the RAUC usecase we do not rely on a checksum anyway as the bundle
signature is checked by RAUC. The only reason to check it that we can
determine if the transferred data is corrupt early, prior to triggering
the RAUC installation.

If the checksum type should be changed in the future, we could either
change the hardcoded style or even make it configurable via the config.
We do not need get_binary_checksum struct for neither case, so replace
it with the calculated checksum string directly.

Signed-off-by: Bastian Krause <bst@pengutronix.de>