hawkbit-client: drop unused md5
The md5 checksum is used nowhere.

For the RAUC usecase we do not rely on a checksum anyway as the bundle
signature is checked by RAUC. The only reason to check it that we can
determine if the transferred data is corrupt early, prior to triggering
the RAUC installation.

Currently the sha1 checksum is used, we can change that in the future
(see the previous commit removing the get_binary_checksum struct). Until
then we should not carry dead code.

Signed-off-by: Bastian Krause <bst@pengutronix.de>