hawkbit-client: refactor get_binary()
Turn struct get_binary into BinaryPayload, add corresponding
binary_payload_free and define it as the autoptr cleanup function. The
use of this struct is incorporated so heavily that is is not possible
easily to separate this from the general get_binary() refactoring.

Further clean and improve the get_binary() function:
- update function documentation
- separate variable declaration and statements
- use g_autoptr
- sanity check input arguments
- use the above mentioned BinaryPayload
- use add_curl_header() to safely add headers without leaking memory in
  error cases
- fix leak by freeing BinaryPayload struct (former get_binary)
- simplify overly complicated if/else constructs

Signed-off-by: Bastian Krause <bst@pengutronix.de>