test: add RAUC D-Bus dummy helper
RAUC requires superuser privileges. This isn't something we want in our
test suite. So add a Python script behaving the same way as RAUC does on
D-Bus without actually installing anything.

This script is meant to be run stand-alone. Due to the nature of
GObject/D-Bus it needs to be run as a subprocess even when used in the
test suite (will be introduced in a future commit).

Signed-off-by: Bastian Krause <bst@pengutronix.de>