config-file: treat auth token and gateway token in config as error
It is surprising that specifying both authentication methods in the
config works, although a warning is emitted. Other code in the
hawkbit-client module then decides to ignore the gateway token in this
case.

Make specifying both authentication methods an error. This is especially
important when more authentication methods (such as client client
certificates) are introduced later.

Signed-off-by: Bastian Krause <bst@pengutronix.de>