hawkbit-client: drop action ID URL regex matching in process_deployment()
It is neither required nor a good idea to get the action ID from an
URL. It is also a bad idea to rely on the "c" GET parameter in
deployment base links. This is an implementation detail which can
change any time in hawkBit. It is not part of the DDI API, rather an
artifact from an example in the API's documentation.

So simply use the action ID from the deployment JSON response and drop
all corresponding regex matching code.

Signed-off-by: Bastian Krause <bst@pengutronix.de>