hawkbit-client: lower log level of incorrect auth tokens in hawkbit_pull_cb()
In a future commit critical warnings will be made fatal. This helps to
find coding issues with the sanity check functions
g_return_if_fail()/g_return_val_if_vail():

  "To make this undefined behaviour visible, if expr evaluates to FALSE,
  the result is usually that a critical message is logged and the
  current function returns."

  https://developer.gnome.org/glib/stable/glib-Warnings-and-Assertions.html#g-return-if-fail

The hawkBit targets might not be registered yet, so this is not
necessarily "critical" in a fatal sense. We will simply try to poll
again with our authentication token.

Signed-off-by: Bastian Krause <bst@pengutronix.de>