rauc-installer: make download thread wait for install thread on run_once
We have a main thread, which starts the download thread if an update is
available for the target. The download thread starts the install thread
once the download succeeded.

For the "run_once" operation mode, the download thread needs to wait for
the install thread to finish (as a first step).

Signed-off-by: Bastian Krause <bst@pengutronix.de>