hawkbit-client: make main loop wait for download thread on run_once
We have a main thread, which starts the download thread if an update is
available for the target. The download thread starts the install thread
once the download succeeded.

For the "run_once" operation mode, the main thread needs to wait for
the download thread to finish (as a second step, see a previous commit
how the download thread waits for the install thread).

Signed-off-by: Bastian Krause <bst@pengutronix.de>