hawkbit-client: use default CURLOPT_WRITEFUNCTION with fd checksum calculation in get_binary()
In order to implement resuming of downloads in a future commit, we need
to be able to calculate checksums on a FILE*.

This allows us to drop the custom curl callback and use the default
CURLOPT_WRITEFUNCTION instead. Now we can drop BinaryPayload
altogether, since filesize/written are now unnecessary, and
FILE*/GChecksum* continue to exist. Furthermore, get_binary()'s
filesize argument could therefor be dropped.

In order still to allow checksum calulcation, add get_file_checksum()
calculating a checksum of the given type on the given FILE*.

Signed-off-by: Bastian Krause <bst@pengutronix.de>