cmake: set PROJECT_VERSION from git
This will set the displayed version number to the output of git
describe and is intended to show the exact version / tag the software
was built from (instead of a hard-coded value).

Note that this requires a tag to be set in order to get results from
`git describe --tags`.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>