hawkbit-client: do not set numerical values to a gboolean
ClientData*->res is a gboolean. TRUE is defined as 1, whereas FALSE is
defined as 0. Setting a numerical value to a gboolean seems wrong in
the first place, but setting it the other way around (0/FALSE on
success and 1/TRUE on error) against intution is confusing.

Thus keep the boolean meaning as long as we are handling a gboolean (in
hawkbit_pull_cb()). Only translate that to an exit code once we are
dealing with a gint in hawkbit_start_service_sync().

Signed-off-by: Bastian Krause <bst@pengutronix.de>