rauc-installer: fix install_context use-after-free on wait
The download thread calls rauc_install() which creates the installation
thread. In "run_once" operation mode, the download thread waits for the
install thread to finish. In this case the download thread reads the
installation's status result from the install_context struct.

The installation thread frees the install_context before the
download thread reads the status result. Fix that by letting the
installation thread know when to leave the freeing of the
install_context struct to the download thread.

Found with address sanitizer (-fsanitize=address).

Fixes: 90fcb8db ("rauc-installer: let rauc_install() return result to allow install->download thread propagation")
Signed-off-by: Bastian Krause <bst@pengutronix.de>