test: conftest: turn bundle_assigned fixture in assign_bundle and reuse it in new bundle_assigned
In the future, we need to parametrize the bundle assignment. So add a
"assign_bundle" fixture which acts just like the previous
"bundle_assigned" fixture, but callable. Now add a new "bundle_assigned"
fixture that simply calls "assign_bundle". This is the most simple case.
More complex cases will follow.

Signed-off-by: Bastian Krause <bst@pengutronix.de>