hawkbit-client: handle deployment's update attribute skip case
If a maintenance window is specifed for a deployment, hawkBit will first
signal a new deployment with ``deployment.update=skip``. The target is
then expected to download the artifact, but skip the actual update for
now. When the maintenance window is "available",
``deployment.update={soft,forced}`` is signalled and the target is
expected to update.

To add support for this, store whether to install the update now or
later in a new Artifact member ``do_install`` and act accordingly.
To keep things simple, follow the usual code flow for delayed bundle
installation and call get_binary(). It will notice that a bundle exists,
do a range request to make sure it is complete and install it
subsequently.

Signed-off-by: Bastian Krause <bst@pengutronix.de>