test: conftest: allow setting number of swmodules/artifacts in assign_bundle fixture
In order to test multiple software modules (also called chunks) and
artifacts, allow passing the number of these to the "assign_bundle"
fixture.

To keep things easy, the fixture sets the software module type as well
as the distribution set type in a way that allows the number of software
modules specified per distribution set.

Signed-off-by: Bastian Krause <bst@pengutronix.de>