introduce config.h
Instead of providing defines as compiler arguments, put them in a
config.h file as autotools and meson would do.

Do not include this by each source file individually but pass it as
compiler -include so that it is not accidentally omitted in new files.

Preparation for possible future adaption/extension of the build system.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>