hawkbit-client: do not pass NULL format to g_strdup_vprintf() in build_api_url()
build_api_url() can be called with NULL to get the base deployment URL.
Passing NULL as a format argument to g_strdup_vprintf() returns NULL if
GLIB_USING_SYSTEM_PRINTF is defined, otherwise it leads to a segmentation
fault. The format argument is explicitly documented as "not nullable",
so we rely on undefined behavior here.

Skip over va_start(), g_strdup_vprintf(), va_end() for path=NULL to fix
that.

Signed-off-by: Bastian Krause <bst@pengutronix.de>