hawkbit-client: move auth header generation to dedicated function
In a future commit, the authentication header will not only be needed by
the download functionality, but also by a function starting the RAUC
streaming installation. This is because RAUC needs to know the
authentication header to pass to hawkBit in order to download the
bundle.

To prepare for that, move the authentication header generation into a
dedicated function that can be called from both code paths.

Signed-off-by: Bastian Krause <bst@pengutronix.de>