test: helper: use basename or python module as logger name
The test suite runs various commands, e.g. rauc-hawkbit-updater, nginx,
but also stand-alone python modules, e.g. rauc_dbus_dummy (see #83,
d63f0e02 ("test: add RAUC D-Bus dummy helper") for the reasoning).

Until now the first component of the command is used as the name of the
logger, so we know which command printed what stdout/stderr lines. Since
the python module is called via
`/path/to/venv/bin/python -m modulename`, the first component might not
be helpful and long.

So use the basename of the first component generally. If this component
happens to be the Python interpreter called with a module, use the
module's name as the name of the logger.

This should help identify the actual command emitting output during
tests.

Signed-off-by: Bastian Krause <bst@pengutronix.de>