test: rauc_dbus_dummy: add remaining DBus properties
When the DBus dummy exposes the dummy interface on the session bus,
generic DBus methods are called to get all the properties.

This results in errors like:

  Exception while handling org.freedesktop.DBus.Properties.GetAll()
  Traceback (most recent call last):
    File "/path/to/rauc-hawkbit-updater/venv/lib/python3.9/site-packages/pydbus/registration.py", line 81, in call_method
      result = method(*parameters, **kwargs)
    File "/path/to/rauc-hawkbit-updater/venv/lib/python3.9/site-packages/pydbus/registration.py", line 110, in GetAll
      ret[local] = GLib.Variant(type, getattr(self.object, local))
  AttributeError: 'Installer' object has no attribute 'Compatible'

Since these errors are not related to rauc-hawkbit-updater, but look
confusing in the test output, implement these properties as stubs.
This makes these errors go away.

Signed-off-by: Bastian Krause <bst@pengutronix.de>