test: read output of process til EOF
The tests run various processes, such as rauc-hawkbit-updater, nginx or
rauc_dbus_dummy. The output of these processes is currently only read
until we see the string/regex we expect. Everything else is not read and
does not end up in the logs.

This is confusing to the person running the tests, because errors,
tracebacks and other output the tests do not rely on is not logged.

To change this, read all relevant process output (until EOF) after the
processes are terminated.

Signed-off-by: Bastian Krause <bst@pengutronix.de>