rauc-installer: pass through auth header for bundle HTTP streaming
The authentication header needs to be passed to the InstallBundle() DBus method
invoking the RAUC streaming installation. In order to do that, add an
`auth_header` argument to rauc_install() and make it pass the header via
the `install_context` struct. The header is finally passed to
InstallBundle() [1] as "http-headers" argument.

For now, `rauc_install()` is called without an actual authentication
(NULL). This will be changed in future commit.

[1] https://rauc.readthedocs.io/en/latest/reference.html#gdbus-method-de-pengutronix-rauc-installer-installbundle

Signed-off-by: Bastian Krause <bst@pengutronix.de>