docs: reference: improve stream_bundle documentation
As explained in #130, bundle streaming with hawkBit does not work work
with hawkBit's default configuration (for most cases). Until now, the
configuration options were only added to the test suite section of the
README. No explanation was provided (not considering the PR and the
commit messages).

To change that, add all required configuration options and an
explanation to the reference docs of `stream_bundle`.

While at it, also mention the default value of the option.

Signed-off-by: Bastian Krause <bst@pengutronix.de>