README/github: bind hawkbit docker container to localhost only
For testing purposes, it is generally not needed to bind against anything
else than localhost. A malicious actor can potentially do bad things with
the default configuration of the hawkbit container, so bind locally only.

Depending on the system configuration, localhost can resolve to ::1 or
127.0.0.1. docker does not support binding to localhost, only to an IP
address. So simply bind to both IPv4 and IPv6.

Signed-off-by: Bastian Krause <bst@pengutronix.de>