README: update configuration
Until now the configuration in the README was a strange mixture of
mandatory and optional options with some oddly specific values not
recommended for field devices.

Let's add all options: optional options are commented and contain either
their default value (if available) or a sane example value.
Sort the options according to the order in `docs/reference.rst`.

Signed-off-by: Bastian Krause <bst@pengutronix.de>